fix: disable auto-shutdown of idle connections
